# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intelligent Anti-Spoof Face Authentication System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterFaceRequest {
  email: string;
  /** @minLength 6 */
  password: string;
  /** 128-dimensional face descriptor vector */
  face_descriptor: number[];
}

export interface RegisterFaceResponse {
  message: string;
  userId: number;
}

export interface LoginFaceRequest {
  email: string;
  /** 128-dimensional face descriptor vector for matching */
  face_descriptor: number[];
  /** Whether the frontend liveness detection passed */
  liveness_passed: boolean;
}

export interface LoginFaceResponse {
  message: string;
  matched: boolean;
  /** Confidence score (0-100), higher is better */
  confidence?: number;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}