# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Intelligent Anti-Spoof Face Authentication System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /register-face:
    post:
      operationId: registerFace
      tags: [auth]
      summary: Register a user with face descriptor
      description: Registers a new user with email, password, and face descriptor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterFaceRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterFaceResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /login-face:
    post:
      operationId: loginFace
      tags: [auth]
      summary: Login with face verification and liveness check
      description: Authenticate a user by matching their face descriptor with liveness validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginFaceRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginFaceResponse"
        "401":
          description: Face not recognized or liveness failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RegisterFaceRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        face_descriptor:
          type: array
          items:
            type: number
          description: 128-dimensional face descriptor vector
      required:
        - email
        - password
        - face_descriptor
    RegisterFaceResponse:
      type: object
      properties:
        message:
          type: string
        userId:
          type: integer
      required:
        - message
        - userId
    LoginFaceRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        face_descriptor:
          type: array
          items:
            type: number
          description: 128-dimensional face descriptor vector for matching
        liveness_passed:
          type: boolean
          description: Whether the frontend liveness detection passed
      required:
        - email
        - face_descriptor
        - liveness_passed
    LoginFaceResponse:
      type: object
      properties:
        message:
          type: string
        matched:
          type: boolean
        confidence:
          type: number
          description: Confidence score (0-100), higher is better
      required:
        - message
        - matched
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error