# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intelligent Anti-Spoof Face Authentication System API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Registers a new user with email, password, and face descriptor
 * @summary Register a user with face descriptor
 */
export const registerFaceBodyPasswordMin = 6;

export const RegisterFaceBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerFaceBodyPasswordMin),
  face_descriptor: zod
    .array(zod.number())
    .describe("128-dimensional face descriptor vector"),
});

/**
 * Authenticate a user by matching their face descriptor with liveness validation
 * @summary Login with face verification and liveness check
 */
export const LoginFaceBody = zod.object({
  email: zod.string().email(),
  face_descriptor: zod
    .array(zod.number())
    .describe("128-dimensional face descriptor vector for matching"),
  liveness_passed: zod
    .boolean()
    .describe("Whether the frontend liveness detection passed"),
});

export const LoginFaceResponse = zod.object({
  message: zod.string(),
  matched: zod.boolean(),
  confidence: zod
    .number()
    .optional()
    .describe("Confidence score (0-100), higher is better"),
});