# lib/api-zod/src/generated/types/loginFaceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intelligent Anti-Spoof Face Authentication System API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginFaceRequest {
  email: string;
  /** 128-dimensional face descriptor vector for matching */
  face_descriptor: number[];
  /** Whether the frontend liveness detection passed */
  liveness_passed: boolean;
}