# lib/api-zod/src/generated/types/loginFaceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intelligent Anti-Spoof Face Authentication System API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginFaceResponse {
  message: string;
  matched: boolean;
  /** Confidence score (0-100), higher is better */
  confidence?: number;
}